[Fix] 앱 정보 뷰 AutoLayout 수정 #129
